Enhancements to DatasetManager
Add ability to create multi-page PDF. Smarter
cleanup of previous analysis (don't delete whole
folder, just the suffixes specified).

+++ cftsdata/util.py
@@ -1,6 +1,8 @@
 import os
 from pathlib import Path
 
+from matplotlib.backends.backend_pdf import PdfPages
+
 
 class CallbackManager:
 
@@ -42,10 +44,11 @@ def get_cb(cb, suffix=None):
 def add_default_options(parser):
     parser.add_argument('folder', type=str, help='Folder containing data')
     parser.add_argument('--reprocess', action='store_true', help='Reprocess all data in folder')
+    parser.add_argument('--halt-on-error', action='store_true', help='Stop on error?')
 
 
 def process_files(folder, glob_pattern, fn, cb='tqdm', reprocess=False,
-                  halt_on_error=True):
+                  halt_on_error=False):
     success = []
     errors = []
     for filename in Path(folder).glob(glob_pattern):
@@ -138,12 +141,20 @@ class DatasetManager:
         filename = self.get_proc_filename(suffix)
         figure.savefig(filename, bbox_inches='tight')
 
+    def save_figs(self, figures, suffix):
+        filename = self.get_proc_filename(suffix).with_suffix('.pdf')
+        with PdfPages(filename) as pdf:
+            for figure in figures:
+                pdf.savefig(figure, bbox_inches='tight')
+
     def save_dataframe(self, df, suffix):
         filename = self.get_proc_filename(suffix)
         df.to_csv(filename)
 
     save_df = save_dataframe
 
-    def clear(self):
-        for filename in self.get_proc_path().iterdir():
-            filename.unlink()
+    def clear(self, suffixes):
+        for suffix in suffixes:
+            filename = self.get_proc_filename(suffix)
+            if filename.exists():
+                filename.unlink()
